requirements.txt
asgiref==3.6.0
click==8.1.3
Django==4.1.7
djangorestframework==3.14.0
face-recognition==1.3.0
face-recognition-models==0.3.0
install==1.3.5
numpy==1.24.2
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
requests
pyTelegramBotAPI==4.10.0
django-debug-toolbar
channels
daphne
opencv-python
channels==3.0.5
imutils==0.5.4
websocket-client==1.8.0
python-dotenv

pyttsx3
websockets
pyaudio
vosk
pveagle
pvrecorder